+++ src/main/java/za/ac/sun/cs/green/service/canonizer/SATCanonizerService.java
@@ -18,6 +18,8 @@ import za.ac.sun.cs.green.util.Reporter;
 import za.ac.sun.cs.green.expr.Constant;
 import za.ac.sun.cs.green.expr.IntConstant;
 import za.ac.sun.cs.green.expr.IntVariable;
+import za.ac.sun.cs.green.expr.IntegerConstant;
+import za.ac.sun.cs.green.expr.IntegerVariable;
 import za.ac.sun.cs.green.expr.Operation;
 import za.ac.sun.cs.green.expr.Variable;
 import za.ac.sun.cs.green.expr.Visitor;
@@ -30,12 +32,15 @@ public class SATCanonizerService extends BasicService {
 	 */
 	private int invocations = 0;
 
+	private long timeTaken = 0;
+
 	public SATCanonizerService(Green solver) {
 		super(solver);
 	}
 
 	@Override
 	public Set<Instance> processRequest(Instance instance) {
+		long start = System.currentTimeMillis();
 		@SuppressWarnings("unchecked")
 		Set<Instance> result = (Set<Instance>) instance.getData(getClass());
 		if (result == null) {
@@ -45,18 +50,19 @@ public class SATCanonizerService extends BasicService {
 			result = Collections.singleton(i);
 			instance.setData(getClass(), result);
 		}
+		timeTaken += (System.currentTimeMillis() - start);
 		return result;
 	}
 
 	@Override
 	public void report(Reporter reporter) {
 		reporter.report(getClass().getSimpleName(), "invocations = " + invocations);
+		reporter.report(getClass().getSimpleName(), "time = " + timeTaken);
 	}
 
-	public Expression canonize(Expression expression,
-			Map<Variable, Variable> map) {
+	public Expression canonize(Expression expression, Map<Variable, Variable> map) {
 		try {
-			log.debug("Before Canonization: {}", expression);
+			//log.debug("Before Canonization: {}", expression);
 			invocations++;
 			OrderingVisitor orderingVisitor = new OrderingVisitor();
 			expression.accept(orderingVisitor);
@@ -65,10 +71,9 @@ public class SATCanonizerService extends BasicService {
 			expression.accept(canonizationVisitor);
 			Expression canonized = canonizationVisitor.getExpression();
 			if (canonized != null) {
-				canonized = new Renamer(map,
-						canonizationVisitor.getVariableSet()).rename(canonized);
+				canonized = new Renamer(map, canonizationVisitor.getVariableSet()).rename(canonized);
 			}
-			log.debug("After Canonization: {}", canonized);
+			//log.debug("After Canonization: {}", canonized);
 			return canonized;
 		} catch (VisitorException x) {
 			log.fatal("encountered an exception -- this should not be happening!", x);
@@ -127,13 +132,10 @@ public class SATCanonizerService extends BasicService {
 			if (nop != null) {
 				Expression r = stack.pop();
 				Expression l = stack.pop();
-				if ((r instanceof IntVariable)
-						&& (l instanceof IntVariable)
-						&& (((IntVariable) r).getName().compareTo(
-								((IntVariable) l).getName()) < 0)) {
+				if ((r instanceof IntVariable) && (l instanceof IntVariable)
+						&& (((IntVariable) r).getName().compareTo(((IntVariable) l).getName()) < 0)) {
 					stack.push(new Operation(nop, r, l));
-				} else if ((r instanceof IntVariable)
-						&& (l instanceof IntConstant)) {
+				} else if ((r instanceof IntVariable) && (l instanceof IntConstant)) {
 					stack.push(new Operation(nop, r, l));
 				} else {
 					stack.push(operation);
@@ -201,7 +203,7 @@ public class SATCanonizerService extends BasicService {
 					}
 				}
 				SortedSet<Expression> newConjuncts = processBounds();
-//				new TreeSet<Expression>();
+				//				new TreeSet<Expression>();
 				Expression c = null;
 				for (Expression e : newConjuncts) {
 					if (e.equals(Operation.FALSE)) {
@@ -209,20 +211,16 @@ public class SATCanonizerService extends BasicService {
 					} else if (e instanceof Operation) {
 						Operation o = (Operation) e;
 						if (o.getOperator() == Operation.Operator.GT) {
-							e = new Operation(Operation.Operator.LT, scale(-1,
-									o.getOperand(0)), o.getOperand(1));
+							e = new Operation(Operation.Operator.LT, scale(-1, o.getOperand(0)), o.getOperand(1));
 						} else if (o.getOperator() == Operation.Operator.GE) {
-							e = new Operation(Operation.Operator.LE, scale(-1,
-									o.getOperand(0)), o.getOperand(1));
+							e = new Operation(Operation.Operator.LE, scale(-1, o.getOperand(0)), o.getOperand(1));
 						}
 						o = (Operation) e;
 						if (o.getOperator() == Operation.Operator.GT) {
-							e = new Operation(Operation.Operator.GE, merge(
-									o.getOperand(0), new IntConstant(-1)),
+							e = new Operation(Operation.Operator.GE, merge(o.getOperand(0), new IntConstant(-1)),
 									o.getOperand(1));
 						} else if (o.getOperator() == Operation.Operator.LT) {
-							e = new Operation(Operation.Operator.LE, merge(
-									o.getOperand(0), new IntConstant(1)),
+							e = new Operation(Operation.Operator.LE, merge(o.getOperand(0), new IntConstant(1)),
 									o.getOperand(1));
 						}
 					}
@@ -353,8 +351,7 @@ public class SATCanonizerService extends BasicService {
 			if (linearInteger && !unsatisfiable) {
 				if (variable instanceof IntVariable) {
 					variableSet.add((IntVariable) variable);
-					stack.push(new Operation(Operation.Operator.MUL, Operation.ONE,
-							variable));
+					stack.push(new Operation(Operation.Operator.MUL, Operation.ONE, variable));
 				} else {
 					stack.clear();
 					linearInteger = false;
@@ -494,9 +491,11 @@ public class SATCanonizerService extends BasicService {
 		private Expression merge(Expression left, Expression right) {
 			Operation l = null;
 			Operation r = null;
-			int s = 0;
+			long s = 0;
 			if (left instanceof IntConstant) {
 				s = ((IntConstant) left).getValue();
+			} else if (left instanceof IntegerConstant) {
+				s = ((IntegerConstant) left).getValue();
 			} else {
 				if (hasRightConstant(left)) {
 					s = getRightConstant(left);
@@ -507,6 +506,8 @@ public class SATCanonizerService extends BasicService {
 			}
 			if (right instanceof IntConstant) {
 				s += ((IntConstant) right).getValue();
+			} else if (right instanceof IntegerConstant) {
+				s += ((IntegerConstant) right).getValue();
 			} else {
 				if (hasRightConstant(right)) {
 					s += getRightConstant(right);
@@ -515,24 +516,24 @@ public class SATCanonizerService extends BasicService {
 					r = (Operation) right;
 				}
 			}
-			SortedMap<Variable, Integer> coefficients = new TreeMap<Variable, Integer>();
-			IntConstant c;
+			SortedMap<Variable, Long> coefficients = new TreeMap<Variable, Long>();
+			IntegerConstant c;
 			Variable v;
-			Integer k;
+			Long k;
 
 			// Collect the coefficients of l
 			if (l != null) {
 				while (l.getOperator() == Operation.Operator.ADD) {
 					Operation o = (Operation) l.getOperand(1);
 					assert (o.getOperator() == Operation.Operator.MUL);
-					c = (IntConstant) o.getOperand(0);
-					v = (IntVariable) o.getOperand(1);
+					c = (IntegerConstant) o.getOperand(0);
+					v = (Variable) o.getOperand(1);
 					coefficients.put(v, c.getValue());
 					l = (Operation) l.getOperand(0);
 				}
 				assert (l.getOperator() == Operation.Operator.MUL);
-				c = (IntConstant) l.getOperand(0);
-				v = (IntVariable) l.getOperand(1);
+				c = (IntegerConstant) l.getOperand(0);
+				v = (Variable) l.getOperand(1);
 				coefficients.put(v, c.getValue());
 			}
 
@@ -541,7 +542,7 @@ public class SATCanonizerService extends BasicService {
 				while (r.getOperator() == Operation.Operator.ADD) {
 					Operation o = (Operation) r.getOperand(1);
 					assert (o.getOperator() == Operation.Operator.MUL);
-					c = (IntConstant) o.getOperand(0);
+					c = (IntegerConstant) o.getOperand(0);
 					v = (IntVariable) o.getOperand(1);
 					k = coefficients.get(v);
 					if (k == null) {
@@ -552,7 +553,7 @@ public class SATCanonizerService extends BasicService {
 					r = (Operation) r.getOperand(0);
 				}
 				assert (r.getOperator() == Operation.Operator.MUL);
-				c = (IntConstant) r.getOperand(0);
+				c = (IntegerConstant) r.getOperand(0);
 				v = (IntVariable) r.getOperand(1);
 				k = coefficients.get(v);
 				if (k == null) {
@@ -563,11 +564,10 @@ public class SATCanonizerService extends BasicService {
 			}
 
 			Expression lr = null;
-			for (Map.Entry<Variable, Integer> e : coefficients.entrySet()) {
-				int coef = e.getValue();
+			for (Map.Entry<Variable, Long> e : coefficients.entrySet()) {
+				long coef = e.getValue();
 				if (coef != 0) {
-					Operation term = new Operation(Operation.Operator.MUL,
-							new IntConstant(coef), e.getKey());
+					Operation term = new Operation(Operation.Operator.MUL, new IntegerConstant(coef), e.getKey());
 					if (lr == null) {
 						lr = term;
 					} else {
@@ -576,22 +576,26 @@ public class SATCanonizerService extends BasicService {
 				}
 			}
 			if ((lr == null) || (lr instanceof IntConstant)) {
-				return new IntConstant(s);
+				return new IntegerConstant(s);
 			} else if (s == 0) {
 				return lr;
 			} else {
-				return new Operation(Operation.Operator.ADD, lr,
-						new IntConstant(s));
+				return new Operation(Operation.Operator.ADD, lr, new IntegerConstant(s));
 			}
 		}
 
 		private boolean hasRightConstant(Expression expression) {
-			return isAddition(expression)
-					&& (getRightExpression(expression) instanceof IntConstant);
+			return isAddition(expression) && ((getRightExpression(expression) instanceof IntConstant)
+					|| (getRightExpression(expression) instanceof IntegerConstant));
 		}
 
-		private int getRightConstant(Expression expression) {
-			return ((IntConstant) getRightExpression(expression)).getValue();
+		private long getRightConstant(Expression expression) {
+			Expression e = getRightExpression(expression);
+			if (e instanceof IntConstant) {
+				return ((IntConstant) e).getValue();
+			} else {
+				return ((IntegerConstant) e).getValue();
+			}
 		}
 
 		private Expression getLeftExpression(Expression expression) {
@@ -615,9 +619,12 @@ public class SATCanonizerService extends BasicService {
 				return Operation.ZERO;
 			}
 			if (expression instanceof IntConstant) {
-				return new IntConstant(factor
-						* ((IntConstant) expression).getValue());
+				return new IntConstant(factor * ((IntConstant) expression).getValue());
 			} else if (expression instanceof IntVariable) {
+				return expression;
+			} else if (expression instanceof IntegerConstant) {
+				return new IntegerConstant(factor * ((IntegerConstant) expression).getValue());
+			} else if (expression instanceof IntegerVariable) {
 				return expression;
 			} else {
 				assert (expression instanceof Operation);
@@ -637,8 +644,7 @@ public class SATCanonizerService extends BasicService {
 
 		private Stack<Expression> stack;
 
-		public Renamer(Map<Variable, Variable> map,
-				SortedSet<IntVariable> variableSet) {
+		public Renamer(Map<Variable, Variable> map, SortedSet<IntVariable> variableSet) {
 			this.map = map;
 			stack = new Stack<Expression>();
 		}
@@ -652,8 +658,7 @@ public class SATCanonizerService extends BasicService {
 		public void postVisit(IntVariable variable) {
 			Variable v = map.get(variable);
 			if (v == null) {
-				v = new IntVariable("v" + map.size(), variable.getLowerBound(),
-						variable.getUpperBound());
+				v = new IntVariable("v" + map.size(), variable.getLowerBound(), variable.getUpperBound());
 				map.put(variable, v);
 			}
 			stack.push(v);